requirements.txt
aiohttp==3.9.3
Flask[async]==3.0.3
numpy
pandas
openpyxl
xlrd
pymilvus==2.4.0
PyMuPDF
python-docx==1.1.0
python-dotenv==1.0.1
tqdm
gunicorn==22.0.0
portalocker==2.10.1
chardet
python-docx
minio
pydantic
elasticsearch==8.15.1